Reorgnanize diagram type/logic shared lib
