+++ process-flow-diagram-component/src/components/Drawer/MenuSidebar.tsx
@@ -1,5 +1,5 @@
 import React, { ChangeEvent, memo, useState } from 'react';
-import { NodeErrors, ParentContainerDimensions, ProcessFlowPart, UserDiagramOptions, processFlowDiagramParts } from '../../../../src/process-flow-types/shared-process-flow-types';
+import { NodeErrors, ParentContainerDimensions, ProcessFlowPart, UserDiagramOptions } from '../../../../src/process-flow-types/shared-process-flow-types';
 import { Box, Button, Chip, Divider, Grid, List, ListItem, ListItemText, Paper, styled, Tab, Tabs, Typography, useTheme } from '@mui/material';
 import ContinuousSlider from './ContinuousSlider';
 import DownloadButton from './DownloadButton';
@@ -13,6 +13,7 @@ import ValidationWindow, { ValidationWindowLocation } from '../Diagram/Validatio
 import NotificationsIcon from '@mui/icons-material/Notifications';
 import { getIsDiagramValid } from '../../validation/Validation';
 import { blue } from '@mui/material/colors';
+import { processFlowDiagramParts } from '../../../../src/process-flow-types/shared-process-flow-logic';
 
 const WaterComponent = styled(Paper)(({ theme, ...props }) => ({
   ...theme.typography.body2,
