+++ src/app/water-process-diagram/water-diagram-connections.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { DiagramIdbService } from '../indexedDb/diagram-idb.service';
-import { ProcessFlowPart, WaterDiagram, WaterProcessComponentType, getNewNode } from '../../process-flow-types/shared-process-flow-types';
+import { ProcessFlowPart, WaterDiagram, WaterProcessComponentType } from '../../process-flow-types/shared-process-flow-types';
 import { AssessmentDbService } from '../indexedDb/assessment-db.service';
 import { Assessment } from '../shared/models/assessment';
 import { Diagram, IntegratedAssessmentDiagram } from '../shared/models/diagram';
@@ -10,6 +10,7 @@ import { firstValueFrom } from 'rxjs';
 import * as _ from 'lodash';
 import { Settings } from '../shared/models/settings';
 import { SettingsDbService } from '../indexedDb/settings-db.service';
+import { getNewNode } from '../../process-flow-types/shared-process-flow-logic';
 
 @Injectable()
 export class WaterDiagramConnectionsService {
