+++ src/app/water/waste-water-treatment/waste-water-treatment-wrapper/waste-water-treatment-wrapper.component.ts
@@ -1,8 +1,9 @@
 import { Component, Input } from '@angular/core';
 import { WaterAssessmentService } from '../../water-assessment.service';
-import { WasteWaterTreatment, WaterAssessment } from '../../../shared/models/water-assessment';
+import { WaterAssessment } from '../../../shared/models/water-assessment';
 import { ConfirmDeleteData } from '../../../shared/confirm-delete-modal/confirmDeleteData';
 import { WasteWaterTreatmentService } from '../waste-water-treatment.service';
+import { WasteWaterTreatment } from '../../../../process-flow-types/shared-process-flow-types';
 
 @Component({
   selector: 'app-waste-water-treatment-wrapper',
