+++ src/app/water/waste-water-treatment/waste-water-treatment.service.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@angular/core';
 import { FormBuilder, FormGroup, UntypedFormGroup } from '@angular/forms';
-import { WasteWaterTreatment, WaterProcessComponent } from '../../shared/models/water-assessment';
-import { getNewProcessComponent } from '../../../process-flow-types/shared-process-flow-types';
+import { WaterProcessComponent } from '../../shared/models/water-assessment';
+import { getNewProcessComponent } from '../../../process-flow-types/shared-process-flow-logic';
+import { WasteWaterTreatment } from '../../../process-flow-types/shared-process-flow-types';
 
 @Injectable()
 export class WasteWaterTreatmentService {
