+++ src/app/water/water-assessment-connections.service.ts
@@ -4,7 +4,7 @@ import { AssessmentDbService } from '../indexedDb/assessment-db.service';
 import { DiagramIdbService } from '../indexedDb/diagram-idb.service';
 import { Assessment } from '../shared/models/assessment';
 import { Diagram } from '../shared/models/diagram';
-import { WaterAssessment, WaterProcessComponent, IntakeSource, WaterUsingSystem, DischargeOutlet, WaterTreatment, WasteWaterTreatment, DiagramWaterSystemFlows, FlowData, KnownLoss, WaterSystemFlows } from '../shared/models/water-assessment';
+import { WaterAssessment, WaterProcessComponent, IntakeSource, WaterUsingSystem, DischargeOutlet, DiagramWaterSystemFlows, FlowData, KnownLoss, WaterSystemFlows } from '../shared/models/water-assessment';
 import { WaterProcessDiagramService } from '../water-process-diagram/water-process-diagram.service';
 import { Settings } from '../shared/models/settings';
 import { WaterAssessmentService } from './water-assessment.service';
@@ -13,7 +13,7 @@ import { WaterSystemComponentService } from './water-system-component.service';
 import { Edge, Node } from '@xyflow/react';
 import { WaterTreatmentService } from './water-treatment/water-treatment.service';
 import { WasteWaterTreatmentService } from './waste-water-treatment/waste-water-treatment.service';
-import { CustomEdgeData, ProcessFlowPart, WaterDiagram } from '../../process-flow-types/shared-process-flow-types';
+import { CustomEdgeData, ProcessFlowPart, WasteWaterTreatment, WaterDiagram, WaterTreatment } from '../../process-flow-types/shared-process-flow-types';
 import { SettingsDbService } from '../indexedDb/settings-db.service';
 
 @Injectable()
