+++ src/app/water/water-assessment.component.ts
@@ -9,8 +9,8 @@ import { Assessment } from '../shared/models/assessment';
 import { WaterAssessmentService } from './water-assessment.service';
 import { ConvertWaterAssessmentService } from './convert-water-assessment.service';
 import { Settings } from '../shared/models/settings';
-import { WasteWaterTreatment, WaterAssessment, WaterProcessComponent } from '../shared/models/water-assessment';
-import { ParentContainerDimensions } from '../../process-flow-types/shared-process-flow-types';
+import { WaterAssessment } from '../shared/models/water-assessment';
+import { ParentContainerDimensions, WasteWaterTreatment } from '../../process-flow-types/shared-process-flow-types';
 import { IntegratedAssessmentDiagram } from '../shared/models/diagram';
 import { WaterAssessmentConnectionsService } from './water-assessment-connections.service';
 import { WasteWaterTreatmentService } from './waste-water-treatment/waste-water-treatment.service';
