+++ src/app/water/water-assessment.service.ts
@@ -2,10 +2,11 @@ import { Injectable } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
 import { ConvertUnitsService } from '../shared/convert-units/convert-units.service';
 import { Settings } from '../shared/models/settings';
-import {  DiagramWaterSystemFlows, DischargeOutlet, IntakeSource, WasteWaterTreatment, WaterAssessment, WaterProcessComponent, WaterTreatment, WaterUsingSystem } from '../shared/models/water-assessment';
-import { ProcessFlowPart, WaterProcessComponentType, getComponentNameFromType, getNewProcessComponent } from '../../process-flow-types/shared-process-flow-types';
+import {  DiagramWaterSystemFlows, DischargeOutlet, IntakeSource, WaterAssessment, WaterProcessComponent, WaterUsingSystem } from '../shared/models/water-assessment';
+import { ProcessFlowPart, WasteWaterTreatment, WaterProcessComponentType, WaterTreatment } from '../../process-flow-types/shared-process-flow-types';
 import { WaterSystemComponentService } from './water-system-component.service';
 import { WaterUsingSystemService } from './water-using-system/water-using-system.service';
+import { getComponentNameFromType } from '../../process-flow-types/shared-process-flow-logic';
 // todo 6875 measur compiler doesn't like pulling in this module because it's from jsx
 
 @Injectable({
